Addressing Henry's comments in the review and also setting this to work every 15 seconds to test. Will turn it back to every 15 minutes after testing.
