+++ ufo/handlers/user.py
@@ -6,13 +6,11 @@ import random
 
 import flask
 from googleapiclient import errors
-from rq import Queue
 
 import ufo
 from ufo.database import models
 from ufo.services import google_directory_service
 from ufo.services import oauth
-import worker
 
 INVITE_CODE_URL_PREFIX = 'https://uproxy.org/connect/#'
 
@@ -323,71 +321,3 @@ def user_toggle_revoked(user_id):
   user.save()
 
   return flask.redirect(flask.url_for('user_details', user_id=user_id))
-
-@ufo.setup_required
-def _check_db_users_against_directory_service():
-  """Checks whether the users currently in the DB are still valid.
-
-  This gets all users in the DB, finds those that match the current domain,
-  and compares them to those found in Google Directory Service for the domain.
-  If a user in the DB is not the domain, then it is presumed to be deleted and
-  will thus be removed from our DB.
-  """
-  db_users = models.User.query.all()
-  config = ufo.get_user_config()
-  credentials = oauth.getSavedCredentials()
-  # TODO this should handle the case where we do not have oauth
-  if not credentials:
-    ufo.app.logger.info('OAuth credentials not set up. Can\'t sync users.')
-
-  try:
-    directory_service = google_directory_service.GoogleDirectoryService(
-        credentials)
-    directory_users = directory_service.GetUsers()
-    for db_user in db_users:
-      # Don't worry about users from another domain since they won't show up.
-      if db_user.domain is not config.domain:
-        continue
-      found = False
-      revoked = False
-      # Search for users based on email field.
-      for directory_user in directory_users:
-        if db_user.email is directory_user['primaryEmail']:
-          found = True
-          revoked = directory_user['suspended']
-          break
-      # Assume deleted if not found, so delete from our db.
-      if not found:
-        if config.user_delete_action is models.CRON_JOB_ACTIONS['delete']:
-          ufo.app.logger.info('User ' + db_user.email + ' was not found in '
-                              'directory service. Deleting from database.')
-          db_user.delete()
-        elif config.user_delete_action is models.CRON_JOB_ACTIONS['revoke']:
-          ufo.app.logger.info('User ' + db_user.email + ' was not found in '
-                              'directory service. Revoking access.')
-          db_user.is_key_revoked = True
-          db_user.did_cron_revoke = True
-          db_user.save()
-        continue
-      if revoked:
-        if config.user_revoke_action is models.CRON_JOB_ACTIONS['delete']:
-          ufo.app.logger.info('User ' + db_user.email + ' was suspended in '
-                              'directory service. Deleting from database.')
-          db_user.delete()
-        elif config.user_revoke_action is models.CRON_JOB_ACTIONS['revoke']:
-          ufo.app.logger.info('User ' + db_user.email + ' was suspended in '
-                              'directory service. Revoking access.')
-          db_user.is_key_revoked = True
-          db_user.did_cron_revoke = True
-          db_user.save()
-        continue
-
-  except errors.HttpError as error:
-    ufo.app.logger.info('Error encountered while requesting users from '
-                        'directory service: ' + str(error))
-
-  def enqueue_cron_user_sync():
-    """Check users currently in the DB are still valid in directory service."""
-    ufo.app.logger.info('Enqueuing cron user sync job.')
-    queue = Queue(connection=worker.CONN)
-    queue.enqueue(_check_db_users_against_directory_service)
